apps/fastapi-backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# HTTP & API
httpx==0.28.1
requests==2.32.3

# AI & ML
openai==1.58.1
anthropic==0.40.0

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Caching & Storage
redis==5.2.1
boto3==1.35.84

# Monitoring & Logging
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2
mypy==1.13.0

# CORS & Middleware (built into FastAPI)
# python-cors - not needed, using FastAPI's built-in CORS

# Environment
python-dotenv==1.0.1

# Geospatial
geopy==2.4.1
shapely==2.0.6

# Background Tasks
celery==5.4.0

# File Upload & Processing
python-magic==0.4.27
pillow==11.0.0

# Email
fastapi-mail==1.4.1

# Rate Limiting
slowapi==0.1.9

# Validation
email-validator==2.2.0

# Stripe for payments
stripe==12.2.0

# Google APIs
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0

# WebSocket support
websockets==14.1